文件序列化至本地保存下载信息

+++ DownloadFile/src/main/java/com/github/downloadfile/helper/DownloadHelper.java
@@ -73,7 +73,7 @@ public class DownloadHelper {
         if(file==null){
             return;
         }
-        if(file.isFile()){
+        if(file.exists()&&file.isFile()){
             file.delete();
         }
     }
